Add TransactionData Interface to OpenID module
