+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/RqesDocumentDigest.kt
@@ -176,7 +176,7 @@ data class RqesDocumentDigestEntry private constructor(
         /**
          * Incompatible version of [at.asitplus.rqes.Method] due to presumably incomplete changes in draft
          */
-        @Deprecated("Unify with [at.asitplus.rqes.Method] asap")
+        @Deprecated("Unify with [at.asitplus.rqes.Method] as soon as new draft allows")
         enum class DocumentAccessMode {
             @SerialName("public")
             PUBLIC,
