+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/Base64URLTransactionDataSerializer.kt
@@ -1,6 +1,6 @@
 package at.asitplus.rqes.serializers
 
-import at.asitplus.rqes.collection_entries.TransactionData
+import at.asitplus.openid.TransactionData
 import at.asitplus.rqes.rdcJsonSerializer
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
@@ -9,7 +9,6 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encodeToString
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
