+++ vck-rqes/src/commonMain/kotlin/at/asitplus/wallet/lib/rqes/helper/OpenIdRqesParameters.kt
@@ -1,6 +1,6 @@
 package at.asitplus.wallet.lib.rqes.helper
 
-import at.asitplus.rqes.collection_entries.TransactionData
+import at.asitplus.openid.TransactionData
 
 /**
  * Wrapper to distinguish RQES related [AuthenticationRequestParameter] members better
