+++ vck-rqes/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/TransactionDataInterop.kt
@@ -2,9 +2,11 @@ package at.asitplus.wallet.lib.rqes
 
 import at.asitplus.dif.InputDescriptor
 import at.asitplus.dif.PresentationDefinition
+import at.asitplus.openid.TransactionData
 import at.asitplus.rqes.QesInputDescriptor
+import at.asitplus.rqes.collection_entries.QCertCreationAcceptance
+import at.asitplus.rqes.collection_entries.QesAuthorization
 import at.asitplus.rqes.collection_entries.RqesDocumentDigestEntry
-import at.asitplus.rqes.collection_entries.TransactionData
 import at.asitplus.rqes.rdcJsonSerializer
 import at.asitplus.rqes.serializers.Base64URLTransactionDataSerializer
 import at.asitplus.signum.indispensable.asn1.KnownOIDs.sha_256
@@ -13,13 +15,18 @@ import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
+import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.util.*
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import kotlinx.serialization.PolymorphicSerializer
-import kotlinx.serialization.Serializer
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.*
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonArray
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.JsonNull
+import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.JsonPrimitive
+import kotlinx.serialization.json.encodeToJsonElement
 
 /**
  * Test vectors taken from "Transaction Data entries as defined in D3.1: UC Specification WP3"
@@ -62,7 +69,7 @@ class TransactionDataInterop : FreeSpec({
         }
     """.trimIndent().replace("\n", "").replace("\r", "").replace(" ", "")
 
-    val transactionDataTest = TransactionData.QCertCreationAcceptance(
+    val transactionDataTest = QCertCreationAcceptance(
         qcTermsConditionsUri = "abc",
         qcHash = "cde".decodeBase64Bytes(),
         qcHashAlgorithmOid = sha_256
@@ -70,11 +77,12 @@ class TransactionDataInterop : FreeSpec({
 
     "Serialization is stable" {
         val encoded = rdcJsonSerializer.encodeToString<TransactionData>(transactionDataTest)
+        encoded shouldContain "type"
         rdcJsonSerializer.decodeFromString<TransactionData>(encoded)
             .shouldBe(transactionDataTest)
     }
 
-    "InputDescriptor serialize" {
+    "InputDescriptor serializable" {
         val input = QesInputDescriptor(
             id = "123",
             transactionData = listOf(transactionDataTest)
@@ -84,7 +92,7 @@ class TransactionDataInterop : FreeSpec({
             .shouldBe(input)
     }
 
-    "TransactionDataEntry.QesAuthorization can be parsed" - {
+    "QesAuthorization can be parsed" - {
         val testVector =
             "ewogICJ0eXBlIjogInFlc19hdXRob3JpemF0aW9uIiwKICAic2lnbmF0dXJlUXVhbGlmaWVyIjogImV1X2VpZGFzX3FlcyIsCiAgImNyZWRlbnRpYWxJRCI6ICJvRW92QzJFSEZpRUZyRHBVeDhtUjBvN3llR0hrMmg3NGIzWHl3a05nQkdvPSIsCiAgImRvY3VtZW50RGlnZXN0cyI6IFsKICAgIHsKICAgICAgImxhYmVsIjogIkV4YW1wbGUgQ29udHJhY3QiLAogICAgICAiaGFzaCI6ICJzVE9nd09tKzQ3NGdGajBxMHgxaVNOc3BLcWJjc2U0SWVpcWxEZy9IV3VJPSIsCiAgICAgICJoYXNoQWxnb3JpdGhtT0lEIjogIjIuMTYuODQwLjEuMTAxLjMuNC4yLjEiLAogICAgICAiZG9jdW1lbnRMb2NhdGlvbl91cmkiOiAiaHR0cHM6Ly9wcm90ZWN0ZWQucnAuZXhhbXBsZS9jb250cmFjdC0wMS5wZGY/dG9rZW49SFM5bmFKS1d3cDkwMWhCY0szNDhJVUhpdUg4Mzc0IiwKICAgICAgImRvY3VtZW50TG9jYXRpb25fbWV0aG9kIjogewogICAgICAgICJkb2N1bWVudF9hY2Nlc3NfbW9kZSI6ICJPVFAiLAogICAgICAgICJvbmVUaW1lUGFzc3dvcmQiOiAibXlGaXJzdFBhc3N3b3JkIgogICAgICB9LAogICAgICAiRFRCUy9SIjogIlZZRGw0b1RlSjVUbUlQQ1hLZFRYMU1TV1JMSTlDS1ljeU1SejZ4bGFHZyIsCiAgICAgICJEVEJTL1JIYXNoQWxnb3JpdGhtT0lEIjogIjIuMTYuODQwLjEuMTAxLjMuNC4yLjEiCiAgICB9CiAgXSwKICAicHJvY2Vzc0lEIjogImVPWjZVd1h5ZUZMSzk4RG81MXgzM2ZtdXY0T3FBejVaYzRsc2hLTnRFZ1E9Igp9"
         val transactionData = rdcJsonSerializer.decodeFromString(
@@ -92,7 +100,7 @@ class TransactionDataInterop : FreeSpec({
             rdcJsonSerializer.encodeToString(testVector)
         )
         "Data classes are deserialized correctly" {
-            transactionData.shouldBeInstanceOf<TransactionData.QesAuthorization>()
+            transactionData.shouldBeInstanceOf<QesAuthorization>()
             transactionData.documentDigests shouldNotBe emptyList<RqesDocumentDigestEntry>()
             transactionData.documentDigests.first().documentLocationMethod shouldNotBe null
             transactionData.documentDigests.first().documentLocationMethod!!.documentAccessMode shouldBe RqesDocumentDigestEntry.DocumentLocationMethod.DocumentAccessMode.OTP
@@ -134,7 +142,7 @@ class TransactionDataInterop : FreeSpec({
         }
     }
 
-    "TransactionDataEntry.QCertCreationAcceptance can be parsed" - {
+    "QCertCreationAcceptance can be parsed" - {
         val testVector =
             "ewogICJ0eXBlIjogInFjZXJ0X2NyZWF0aW9uX2FjY2VwdGFuY2UiLAogICJRQ190ZXJtc19jb25kaXRpb25zX3VyaSI6ICJodHRwczovL2V4YW1wbGUuY29tL3RvcyIsCiAgIlFDX2hhc2giOiAia1hBZ3dEY2RBZTNvYnhwbzhVb0RrQytEK2I3T0NyRG84SU9HWmpTWDgvTT0iLAogICJRQ19oYXNoQWxnb3JpdGhtT0lEIjogIjIuMTYuODQwLjEuMTAxLjMuNC4yLjEiCn0="
 
@@ -143,7 +151,7 @@ class TransactionDataInterop : FreeSpec({
             rdcJsonSerializer.encodeToString(testVector)
         )
         "Data classes are deserialized correctly" {
-            transactionData.shouldBeInstanceOf<TransactionData.QCertCreationAcceptance>()
+            transactionData.shouldBeInstanceOf<QCertCreationAcceptance>()
         }
 
         "Encoding-Decoding is correct" {
@@ -161,7 +169,6 @@ class TransactionDataInterop : FreeSpec({
             rdcJsonSerializer.decodeFromString<PresentationDefinition>(presentationDefinitionAsJsonString)
         val first = presentationDefinition.inputDescriptors.first()
             .shouldBeInstanceOf<QesInputDescriptor>()
-        @Suppress("DEPRECATION")
         first.transactionData shouldNotBe null
     }
 })
